Added callback tests
- callback_test() checks if the callback is executed if, and only if the
condition is satisfied

- callback_in_thread() tries to look for problems when accessing the
static variable from multiple threads

- Made the Value::new() function const, to be able to initialize static
variables with it
